pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "impact-engine-allocate"
version = "0.1.0"
description = "Minimax regret portfolio optimization for the impact engine pipeline."
authors = [
  { name = "eisenhauerIO", email = "info@eisenhauer.io" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pulp",
  "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest", "nbmake", "ruff", "pre-commit"]
notebooks = ["jupyterlab", "matplotlib", "pandas", "numpy"]

[tool.hatch.build.targets.wheel]
packages = ["impact_engine_allocate"]

[tool.hatch.envs.default]
features = ["dev", "notebooks"]

[tool.hatch.envs.default.scripts]
test = "pytest -v --nbmake {args}"
lint = "ruff check . && ruff format --check ."
format = "ruff format ."

[tool.hatch.envs.docs]
dependencies = [
  "sphinx>=7.0",
  "sphinx-rtd-theme",
  "myst-parser",
  "nbsphinx",
  "pypandoc",
  "ipykernel",
  "nbconvert",
  "pulp",
  "pyyaml",
  "matplotlib",
  "pandas",
  "numpy",
]

[tool.hatch.envs.docs.scripts]
build = "sphinx-build -b html docs/source docs/build/html"

[tool.ruff]
extend-exclude = ["_external", "docs/build"]
line-length = 120

[tool.ruff.lint]
select = ["D", "E", "F", "I"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["D", "E402", "F", "I"]
"tests/**" = ["D"]
"**/tests/**" = ["D"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests", "docs/source"]
